Start working on GUI for machine-check
I decided to use the combination of the wry and tao packages, which allows using browser-based rendering and is fairly lightweight (2 MB in Windows, release mode). This should allow for relatively painless GUI integration, with dry compile-time (~30 s) as the most problematic part (a no-gui build option needs to be investigated).

Only displays a static page in a stand-alone executable for now.