Split verification into separate steps
This will allow viewing the progress of verification in the GUI as refinement occurs. The question was where to put the point where the loop splits. I chose to put it after model-checking and start the framework with an empty state space, which ensures the verification results are only returned in the verification step and the framework should be quickly constructable (initial state space is constructed only afterwards). However, this means that the framework is "invalid" before the first verification step, so care must be taken.