Add start of machine-check-exec control by GUI
Currently just simple printing of edges and a button to go to next verification step. RwLock used for synchronisation on the backend side, needs to be thought out on the GUI side. Reset needs to be thought out, the way of reconstructing the framework does not play well with closures (take_mut or optionals would be needed). After that, it will be time to expose more data and visualise nicely.