Make the incoming and outgoing references nicer
The new reference style is akin to the labels in hardware EDA tools such as KiCad to be familiar to people.

I also added handling of a special case when the node has only one successor and the successor has incoming references. In that case, there is no need to offset the successor to make space for the references.