Add a quick-and-dirty favicon
Providing a favicon suppresses errors in browser console and on the command-line. Interestingly, even though the favicon is requested by the browser, it is not used by the window. Added a bit of code to load it without requiring another package. The favicon is own work, and is just a quick-and dirty ten-point star (to signify a machine cog) with orange fill (to signify machine-check is written in Rust). It can be replaced with something better once there is time.