Add building of WebAssembly frontend for GUI that will replace JS
This will allow faster development and better integration of the GUI backend and frontend. It also allows for a future option of tools based on machine-check running fully in the browser including the GUI. However, there is added complexity of needing the wasm32-unknown-unknown toolchain and wasm-bindgen. Since the GUI is an opt-in feature, this feels acceptable to me.