Update node IDs to be 64-bit instead of 128-bit
I previously updated the node IDs to be 128-bit instead of usize-sized in 140a370160de6a759e2b1b3b55c0acf8ad364f23 . This was due to the realisation that since we are garbage-collecting states, the node IDs might overflow on 32-bit systems without first exhausting memory. However, it is sufficient to use IDs fixed to be 64-bit due to the sheer generation speed necessary to overflow them.