Change GUI frontend/backend comms from JSON to MessagePack
This should be a bit faster and, much more importantly, allow for using u64 values such as node IDs. It would be the most preferable to use a completely binary exchange without encoding e.g. field names, but it would require more care and be more prone to problems than the Serde approach, so I will leave that for future.