Fix GUI build.rs so the frontend build does not occur always
The commit 04dbd9cc50a6d1a53bb90df9e07ce028535b45cd fixed the frontend not rebuilding when patched dependency paths were changed, but made the rerun-if-changed include the GUI package directory, which conventionally includes the target directory, so the frontend was always rebuilt. Here, I added some logic to filter the patches so that rerun-if-changed only includes paths for actual dependencies.