Rework the GUI backend synchronisation
Now uses a single backend worker thread and message passing for a hopefully more rugged functionality.

Previously, there were two mutexes, one for stats intended for short locking and one for the workspace intended for long locking during computation, with computations that ran asynchronously with respect to the request from the frontend spawning a new thread.

Now, each server thread responding to a request from the frontend tries to pass a message on a rendezvous synchronous channel: if it is unable to, it means the worker is busy. If it is able to, it also passes a message-sender back to the server thread on which it waits, so the server is able to do some synchronous computations and then pass the response back to the server thread. It can then work further without blocking the frontend. There is now only one RwLock for the stats, and deadlocks should be (hopefully) prevented. There may be some theoretical possibility for starvation due to the non-priority RwLock, but this should not occur in practice.