Introduce WASM frontend preparations so the users do not need to build it
Building the WebAssembly GUI frontend is awkward to arrange. Fortunately, we can spare the users this effort by pre-building the WebAssembly and only checking that the built package hash matches the sources on the user's machine.

This has minor source-code-inspection-mindset implications as it is no longer as possible to audit the built WebAssembly, but the affected area is minimal: the frontend runs in a browser with only a custom protocol channel between it and the backend, which is still fully auditable.