Fix WASM frontend preparations so it is possible to publish on crates.io
For publishing, the build script cannot make any filesystem changes except for in OUT_DIR, so I changed my angle: now, the publishing run just creates the frontend package below OUT_DIR, computes its hash, and compares to the built wasm hash. This is only intended for ensuring the interaction between the frontend and backend is not broken, not for security reasons.

Also, cargo publish rewrites Cargo.toml, so I removed that (and the created .gitignore) from hash consideration.