Post frontend request content in a header instead of body for compatibility
This fixes a problem I encountered in Linux (Lubuntu 24.04): the request body sent to the backend was empty. Sending the data in a header is a bit of a kludge, but it works well, and the data sent from frontend to the backend are not expected to be large, so there should be no problem.