Fix building GUI frontend
There were problems with building the GUI frontend: the content/wasm directory was usually not removed, yet the created machine-check-gui-wasm package was removed.

The behaviour should be fixed now: the machine-check-gui-wasm package will not be removed unless building when preparing, and the content/wasm directory will always be removed when building.

This should make sure that the target files for build speedup are kept yet an error is emitted when the build fails.