Rework property parsing and display (COMPATIBILITY BREAK)
Thinking about it, it will be the best to have the properties specified in a proper subset of Rust, so they can be subsequently parsed by syn and processed by normal translation, with limited interpretation capabilities.

I therefore reworked the parsing of temporal properties: now, they are specified like Rust macro invocations, i.e. instead of A[a]U[b], we have AU![a,b], and instead of EG[a], we have EG![a]. The bang at the end makes things nicely unambiguous and removes collisions with state variable identifiers.

I also fixed a problem where the nested and/or had no parentheses displayed, which could be ambiguous (or even misleading when assuming they have precedence as in Rust).