Fix non-monotonicity bug
There was a non-trivial bug where monotonicity was lost: precision of a given state could have been more precise than the precision of a state that it fully covers. This violated the monotonicity requirement when a refinement moved to the covered state. I have fixed this for now by iterating over the precision map and fixing the coverage, although a faster algorithm should be implemented in the future.