Fix introduced bug where state space generation would infinitely loop
This seems to be introduced in bea9a9a584d2c5b2e28638ba275105b18616599b: while the logic was switched to queuing the states if their successors were empty, this no longer serves to ensure that the generation is done only once for the newly introduced state.

I added the logic that checks that the current node has no direct successor, which should now work as intended.