Rework and re-enable state space garbage collection
The solution is still a bit uglish: while a panic ensures if we are getting the data for a state id that is no longer valid, it is possible to assign a new state id to the same data. It would probably be better to switch to Arcs to both the state data and id inside the framework, but that would have different problems of its own. I need to think about it more.