Rework the args conflicts
The "gui" argument should not conflict with "inherent", as that is the default behaviour of the GUI: start with just the inherent property available for verification. On the other hand, the "property" should conflict with "inherent" as we are either ultimately trying to verify the given property or the inherent property.