Add tests for machine-check-avr
This is based on the programs I have crafted for benchmarking with the whole artefact available on Zenodo at https://zenodo.org/records/15109092 . I have added only the original source and hex files to this repository to avoid polluting it. This will enable easier testing that new features do not introduce bugs or preclude verification of previously verifiable programs in reasonable time.