Add subsuming blocks after resolving scopes
Normal code blocks are meaningless except for scoping variables in the supported subset of Rust for translation, so they can be subsumed in the outside block by flattening them. This will allow for a bit simpler intermediate representation.

The current plan is to refactor the translation so that custom simpler IR will be used instead of the syn AST. This will in turn allow further improvements to translation and types.