Correct the typing to have PhiArg and PanicResult types above references
The types are now composed of basic types (bitvectors, arrays, Booleans) that can be referenced once, and they can then be put into PhiArg or PanicResult which are specific to the translation. This is fairly simple but can be powerful.

This commit should complete the type inference refactor to WIR, so we can move to other parts of the translation.