Rewrite conversion to mck types to use WIR
This necessitated another generic for the fundamental type, as the distinction between bitvector/unsigned/signed is erased in this stage.

This is a straightforward rewrite to use WIR, it will probably be for the best to detect the special machine_check and std function calls early and encode that information into WIR. Then, type inference nor this stage will use special path detection, which is ugly and duplicates code.