Encapsulate WIR ident so the manipulation is easily seen
This should help integrating normalisation of scoping and constructs into the WIR folding. It seems to me that the panic conversion should be probably sandwiched between WIR folding and conversion to SSA, as it would be quite awkward to do it afterward. The type inference should probably remain after SSA, to be able to infer nicely in the future.