Move checking supported paths from visitor to WIR from syn
For future support of method calls, we should make sure no traits are imported into the module scope so we can resolve methods unambiguously. This means use declarations should be checked as well. Since it seems impossible to allow use of both types and functions but not traits without outside information, it seems that it will be the best to be able to import just types within machine_description, by emitting code that does not compile if the use declaration does not name a type.