Add dual-interval domain stub
The dual-interval domain came to me when studying ways to make an interval domain that can handle combinations of unsigned and signed operations nicely, and forms a lattice unlike the approach from Signedness Ignorance is Bliss ( https://dl.acm.org/doi/10.1145/2651360 ). The resources needed are slightly higher, but the abstraction computation is in fact not currently very costly compared to model-checking and refinement.

This is not yet scientifically published, so I hope I am not making a mistake in my reasoning and everything works out.