Implement combined interval refinement marking
This is currently just doing the standard three-valued bit-vector marking, and should suffice for now until better properties are implemented because it interacts with them.

The combined domain seems to have some bug(s) now as it does not pass tests with a bad assert, but it properly builds and passes for some simpler examples.