Fix dual-interval bugs
There were two bugs in my initial implementation: the phi function wrongly did not consider a half present in just one of the two operands, and the creation of a three-valued function from an interval did not use a proper mask (all bits down from the first unknown position including it).

Now, all tests pass with dual-interval, although the programs take A LOT of time and computation power. This is better than expected: the combined domain is more precise and there is nothing to counteract this precision, so the abstract state spaces should be larger. The advantage of the domain should properly show when making refinements more clever, which is currently blocked by property handling.