Add initial support for showing the dual-interval domain values in the GUI
One thing that does not seem to have a particularly great solution in Rust is that we want the versions of bit-vectors generic with respect to the bit-vector length for speed and type-checking purposes, but in other places, we want the width to be specified at runtime for better serialisation (and, in the near future, interpretation).

After long thinking, I decided to have the domain printing in constructed field values, which probably eventually will be extended to support the operations as with the generic bit-vectors. I will make this a bit nicer as well.