Allow GUI and dual-interval domain in machine-check-hw
There is still some wonkiness: the preparation and used features should be the same, but this is currently not checked. It would also be nice to cleverly patch the package paths as currently, the build fetches the packages from outside (crates.io). This is nice-to-have, but I feel this is a good place to be for now.