Fix wrong type of fixed-point operators converted from CTL
The greatest/least fixed point choice was done wrong (even though I described the correct way to do it in a comment). It is necessary to consider the temporal quantifier, not the path quantifier, which is handled in EX/AX only.

Essentially, for eventually/until, we search for an actual state that fulfills this, so we start with everything false and search for true, i.e. search for the least fixed point. For globally/release, we start with everything true and search for something that blocks them, i.e. search for false, which is the greatest fixed point search.