Make EX and AX work dually to each other
This removes the weird negation usage. Everything now works as expected from mu-calculus checking. The performance is still very bad compared to CTL, but should now be easily enhanced as the logic is much simpler.

There is also the failing programs_2025_04::toy::avr_toy_blink_debug_invariant_lock test, which needs to be investigated as well.