Add preparation for incremental model-checking
The idea is that to preserve the order of fixed-point computation when checking incrementally, we must preserve the order of considering the next states. We do this by considering "waves" of the fixed-point operator iteration: the first wave computes the atomic propositions and constant, the second one is applied to that, etc. More thinking needs to be done for nested operators.

Otherwise, this commit ensures that the dirty states are always reset to the ground value when starting fixed-point computation, as consistent with mu-calculus.