(BROKEN) Work on incremental model-checking with history
Incremental model-checking of mu-calculus has been a much bigger problem than anticipated due to the interplay of the next operators which can interact with clean states and fixed-point operators which reset the variables. It seems necessary to keep the full history to be able to only partially recompute, but the exact correct way of doing it is really problematic.