(MAYBE BROKEN) Rewrite mu-calculus checking to use fixpoint history
Previous approaches did not yield correct and performant algorithms for incremental mu-calculus model checking. I think this was due to the difficult juggling of current labellings, prone to errors when history is introduced.

In this commit, the model-checking is reworked so that only the fixpoint history is retained. Everything else is recomputed on demand. All states are currently recomputed, so the performance is not good enough to compute all tests, but the new code seems much more elegant than previously and uses history as a first-class (and only, with the exception of current time) citizen.