(MAYBE BROKEN) Add slow but seemingly correct incremental checking
The idea is to invalidate the model-checking and do it from the ground up if the timespans of fixed-point computations do not match. This is still slow as fixed-point calmness is currently not implemented and there is no slack in the times, which may lead to a lot of invalidations.

I also redid the time computation as the previous was problematic: now, each new fixed-point iteration increments the time BEFORE the iteration is computed, including the last iteration where the increment is no longer skipped. This also means that the new fixed-point variable values take the latest time value BEFORE the current one, not up to the current one as previously.