(MAYBE BROKEN) Move figuring out if the property is closed-form from model-checking
Under the current model-checking, we know that a subproperty can be "calm" (and thus not recomputed) if it is in closed-form, i.e. contains no variables that are not quantified by the fixed-point quantifiers. Being in closed-form is an intrinsic property of the (sub)property, so I have moved it.

I also moved the test for calmness, so it should not push new computations before leaving, which seems a bit better, but the computation hijinks should probably be all moved to one place at the end.

Also reworked the way fixed points are computed for more clarity.