(MAYBE BROKEN) Fix problems in incremental model checking
There were two problems. First, the fixed points actually must be updated when there are old values, otherwise there is a chance of missing out on updates. The other problem was that it is necessary to consider the states that can be reached within one time instant, which is determined by the maximal transition depth (number of nested next operators).

Now, all normal tests (slowly) pass, but normally ignored tests (calibration) are still much too slow.