(MAYBE BROKEN) )Add model-check time padding and opt. double-checking
Things are getting good! The timings now seem to work well and can be padded to avoid frequent invalidations.

The double-checking capability can be turned on by an environment variable, gives a warning when on, and checks that the incremental history and computations are the same as non-incremental. This needs to be modified further to squash the timings beforehand, ensuring that the double-checking succeeds even when the original end times differ (which is typical).

Some retouching is also needed, and we can gain the performance by restricting the considered states. This should hopefully get us somewhere reasonable for passing all tests.