(MAYBE BROKEN) Remove the need to consider all preceding states
In model-check.rs, there was a provision that forced setting all (even non-direct) predecessors as dirty. This is problematic as state spaces get larger. I have changed this to just consider the predecessors that can be reached in a single iteration. It is newly necessary to correctly handle the cases of the predecessors no longer changing as they did previously, which should be handled now. Things can be sped up now.