Retouch the mu-calculus computation code
I added iterating over the smaller colection during fixed-point iteration when considering the previously updated states that also have to be backward-affected. This is done using a custom iterator for code deduplication. Unfortunately, it seems that deduplication of the computation of forward- and backward-affected states without performance loss does not seem to be possible that nicely, since we cannot select an iterator impl.