Start implementing new property interpretation
The intent is to make property and description processing almost the same, based on normal Rust, so both will be able to make use of interpretation. This will make it possible to speed up compilation and runtime, eliminate the compilation stack size issues, and easily add relational and other domains.

The new checking is under construction and currently disabled in this commit.