Add parsing of fixed-point and unary CTL operators in new properties
The property parsing is currently incredibly uglish. The integration with interpretation is almost done, but time is pressing me. Perhaps it will be for the best to pause the progress on interpretation and instead focus on parametric verification (which should be very easy to add now) and ensuring everything works for now.