Remove the start of new property rewrite
This will make it possible to finish the release with the properties enhanced to mu-calculus, but without changing the properties to the Rust description processing and intepretation, so that the current release can be finished in August. The properties change is currently retained in the branch new_property and will be merged to the dev branch after the release.