Implement parametric model-checking
This changes the behaviour of computation of next states in mu-calculus. No other change was necessary other than fixing the behaviour of state space insertion and removal, where the parameter sets were not properly unionised in the initial implementation.

I added an example for parametric model-checking and things look fairly nice, although the performance will probably be bad as of now for many things where we would want to use this due to the lack of suitable domains.