Fix squashing of computations for double-checking
I did some funky subtractions previously, but this does not always work. The new solution is simple: just make a map of original computation times where there is something happening in the histories pointing to the squashed times, then choose the squashed time for the original (start/end) time by looking for  the entry from (inclusive) the original time upwards.

There is still some problem with histories when checking using the GUI, I am on it.