Make the property icons in the GUI a bit better visible
It will be better to replace the builtin Unicode icons that depend on the browser and the operating system with proper custom icons that will look good everywhere. Unfortunately, I tried some icon libraries and I am blocked by the old version of Rust that is supported by machine-check, so this will have to wait until another version.