Change non-monotone property error to proceed syntactically
The runtime version actually did not work well for alternating quantifiers since the more nested quantifiers can make the outside ones proceed "non-monotonically" with respect to the iterations. Syntactically ensuring that there are only double negations of variables wrt. the fixed-point operators is an easy fix for now, although it might be better to think up something non-syntactic in the future.