Add tests of examples and book examples
This should prevent unexpected regressions and changes in refinement behaviour. This introduces duplication, but it would be hard to prevent it. Something to think about.

I found that the example in mu-calculus does not correspond to the property AFG, and the correct one seems to die with an assertion failure. Bugfix is coming.

For simple_risc, it is actually better not to use decay for verifying one of the properties, I removed the comment to use it there.