(MAYBE BROKEN) Fix incremental checking without calmness
There were two problems. First, we must take look at backward-affected states that have been recomputed both in previous and current time, or there might be problems with non-updates. Second, we have to marry wanting to update (and make dirty) when the values are not the same as previously with the need not to overwrite the previous unknown reason. As such, we only overwrite unknown reason when it happened at the same time as current update.

Also fixed the double-checking not filtering states no longer in state space from fresh property checker, leading to spurious panics.