Fix incremental computation problem uncovered by calmness
The remaining problem was that the insertion returned when the value was already implied by previous, but did not remove things at the exact time.

All tests pass now (although fairly slowly, in 15 minutes on Core 12900), and mu-calculus infinitely often also does. Presumably due to the combination of larger reasons (should be easily mitigated by not storing any trivial choices)  and more computation due to computing both when current and previous updated (not easily mitigated).