(BROKEN) Start deduplicating runtime and compile domains
This will allow using both runtime and compile-time bitvector widths without any duplicate code. Should be a zero-cost abstraction for compile-time code.

The idea is that instead of the actual widths generics, we have a generic structure that either retains the width using generics or using a variable.

Heavily broken for now.