Fix GUI not working with features
The problem was that different features could produce different versions of AbstractValue and Conclusion from the ones used when compiling frontend, which is compiled without features.

The solution was to convert Conclusion to ParamValuation and AbstractValue to newly added display structs. This actually allows quite nicely composing the domain display.