Update wry and tao to solve Windows GUI link errors
Annoyingly, there are some link errors on Windows 11 with wry 0.50.5, which is automatically used if not locked to 0.50.4. This makes it impossible to compile the GUI in previous versions when this problem occurs. Probably no good solution except for hoping that the updated packages will not have this problem in the future.