Fix writing to arrays in nested ifs
This should have worked, but the systems including it did not compile. The problem is that when we abstract the control flow, we now have both an if and an else part, and we cannot pass values of non-Copy types to both at the same time. It is necessary to add cloning when this is detected.