Add RISC-V tests
For the timed example, only a few tests where the result is given in reasonable time for a test (at most tens of seconds) are specified. There is a lot that could be done performance-wise; for spinlocks such as the one in the timed example, some kind of smart decay will be necessary.

All tests pass.